Adding utils.py
