+++ bot/data/nlu_interactive.md
@@ -1,0 +1,12 @@
+## intent:choose
+- [2](choice)
+- [3](choice)
+
+## intent:greet
+- oi
+
+## intent:inform
+- como está o clima em [brasilia](locale)?
+
+## intent:sports
+- me fala um esporte para [brasilia](locale)?
