+++ bot/domain.yml
@@ -1,47 +1,48 @@
 %YAML 1.1
 ---
 actions:
-- action_humidity
-- action_local
-- utter_greet1
-- utter_good_morning
-- action_temperature
-- utter_goodbye
-- action_sports
-- utter_elaborate
-- utter_gaia_error
-- utter_help
-- action_sunrise_sunset
-- utter_greet2
-- utter_greet
-- utter_change_subject
-- utter_good_night
-- utter_out_of_scope
-- action_specific_sport
 - action_wind
-- utter_how_are_u
-- utter_hurricane
-- action_pressure
-- utter_compliments
-- utter_default
-- utter_how_do_u_know
 - utter_good_afternoon
-- utter_who_are_u
-- action_sky
+- action_temperature
+- utter_change_subject
+- action_sunrise_sunset
+- utter_greet1
+- utter_greet
+- action_pressure
+- utter_good_morning
+- utter_greet2
 - utter_user_error
+- utter_goodbye
+- utter_hurricane
+- action_humidity
+- action_sports
+- utter_help
+- utter_compliments
+- action_specific_sport
+- utter_who_are_u
 - utter_anything_else
+- utter_out_of_scope
+- utter_how_are_u
 - action_weather
+- utter_elaborate
+- action_local
+- utter_default
+- action_sky
+- utter_good_night
+- utter_how_do_u_know
+- utter_gaia_error
 entities:
-- sport
 - locale
-- hurricane
-- choice
 - type
+- hurricane
+- sport
+- choice
 intents:
-- specific_sport
-- greet
+- inform
 - choose
 - sports
+- greet
+- specific_sport
 - goodbye
 - sky
 - humidity
@@ -55,7 +56,6 @@ intents:
 - change_subject
 - how_are_u
 - compliments
-- inform
 - hurricane
 - elaborate
 - locale
